# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Office Works",
    "description": "Production-grade FastAPI backend with auth (JWT+OAuth+RBAC) and LLM chat proxy domains.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register a new user",
        "description": "Register a new user account and send a verification email.\n\nFastAPI validates and normalizes ``payload`` before this handler delegates\nregistration to the auth application service.  The response does NOT\ninclude a JWT — the user must verify their email first (or the calling\nclient skips verification in non-prod envs).\n\nRaises\n------\n409\n    If a user with the given email already exists.",
        "operationId": "signup_api_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-email/{token}": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify email address",
        "description": "Mark a user's email as verified using the token from the email link.",
        "operationId": "verify_email_api_v1_auth_verify_email__token__post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login and receive JWT pair",
        "description": "Authenticate with email + password and receive an access + refresh JWT pair.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Rotate refresh token",
        "description": "Exchange a refresh token for a new access + refresh token pair.\n\nThe old refresh token is immediately revoked.  Reuse of an already-rotated\ntoken triggers family revocation (all sessions for the user are terminated).\n\nRaises\n------\n401\n    If the token is invalid, expired, revoked, or a reuse attack is detected.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout and revoke tokens",
        "description": "Revoke the refresh token and blacklist the current access token.\n\nThe client should discard both tokens after calling this endpoint.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/password-reset": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request password-reset email",
        "description": "Send a password-reset link to the given email.\n\nAlways returns the same 202 response regardless of whether the email exists\n(prevents user enumeration).",
        "operationId": "request_password_reset_api_v1_auth_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password-reset/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Apply password-reset token",
        "description": "Apply a password-reset token and set a new password.",
        "operationId": "confirm_password_reset_api_v1_auth_password_reset_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Current authenticated user",
        "description": "Return the profile of the currently authenticated user.",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/domains__auth__schemas__auth_schemas__UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/oauth/{provider}/login": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "OAuth2 authorization URL",
        "description": "Generate an OAuth2 authorization URL for the given provider.\n\nThe returned ``state`` nonce is stored in Redis for CSRF validation in the\ncallback.  Supported providers: google,kakao,naver,microsoft.\n\nRaises\n------\n400\n    If *provider* is not configured or unsupported.",
        "operationId": "oauth_login_api_v1_auth_oauth__provider__login_get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthLoginURLResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/oauth/{provider}/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "OAuth2 callback — exchange code and redirect to frontend",
        "description": "Complete the OAuth2 flow and hand off the app JWT pair to the SPA.\n\nValidates the ``state`` nonce against Redis to prevent CSRF attacks, then\non success redirects (302) the browser to\n``{frontend_url}/auth/callback#access_token=…&refresh_token=…``.  On any\nfailure (invalid state, provider/exchange error, provisioning error) it\nredirects to ``{frontend_url}/login?error=oauth``.",
        "operationId": "oauth_callback_api_v1_auth_oauth__provider__callback_get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Authorization code from provider",
              "title": "Code"
            },
            "description": "Authorization code from provider"
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "CSRF state nonce",
              "title": "State"
            },
            "description": "CSRF state nonce"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Request body for POST /auth/login."
      },
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "LogoutRequest",
        "description": "Request body for POST /auth/logout."
      },
...
      "OAuthLoginURLResponse": {
        "properties": {
          "authorization_url": {
            "type": "string",
            "title": "Authorization Url"
          },
          "state": {
            "type": "string",
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "authorization_url",
          "state"
        ],
        "title": "OAuthLoginURLResponse",
        "description": "Response body for GET /auth/oauth/{provider}/login."
      },
      "PasswordResetConfirmRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "token",
          "new_password"
        ],
        "title": "PasswordResetConfirmRequest",
        "description": "Request body for POST /auth/password-reset/confirm."
      },
      "PasswordResetConfirmResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Password reset successfully."
          }
        },
        "type": "object",
        "title": "PasswordResetConfirmResponse",
        "description": "Response body for POST /auth/password-reset/confirm."
      },
      "PasswordResetRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "PasswordResetRequest",
        "description": "Request body for POST /auth/password-reset (request reset link)."
      },
      "PasswordResetRequestResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "If an account with that email exists, a reset link has been sent."
          }
        },
        "type": "object",
        "title": "PasswordResetRequestResponse",
        "description": "Response body for POST /auth/password-reset.\n\nThe message is intentionally generic to prevent account enumeration."
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "description": "Request body for POST /auth/refresh."
      },
...
      "SignupRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "display_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "title": "SignupRequest",
        "description": "Request body for POST /auth/signup."
      },
      "SignupResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/domains__auth__schemas__auth_schemas__UserResponse"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Verification email sent."
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "SignupResponse",
        "description": "Response body for POST /auth/signup."
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "minLength": 1,
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Expires In",
            "description": "Access token TTL in seconds."
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "description": "JWT pair returned by login and refresh endpoints."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyEmailResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Email verified successfully."
          },
          "user": {
            "$ref": "#/components/schemas/domains__auth__schemas__auth_schemas__UserResponse"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "VerifyEmailResponse",
        "description": "Response body for POST /auth/verify-email/{token}."
      },
...
      "domains__auth__schemas__auth_schemas__UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "is_verified",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "Public user representation — never includes hashed_password."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}